refactor: rename `EntityRegistrationForm` steps based on context
