+++ src/domains/transaction/components/EntityRegistrationForm/FormStep.tsx
