+++ src/domains/transaction/components/EntityRegistrationForm/ReviewStep.tsx
